Add Saeid's cnn and try to make it work.
model-2.ipynb .. is self explanatory. We're just trying to make it work
by preparing some data to give it.

cnn1.py contains the CNN and a simple wrapper function that helped with
the initial experiments but is overlooking a lot of things

the code has been sligthly modified to run and should be probably
polished by the original author :D